refactor: drop leading underscores from module and class names
Rename the leading-underscore modules to satisfy the no-`_`-prefix file rule:
_context_serialization.py -> context_serialization.py, _sentinel.py ->
sentinel.py, _dev_bridge.py -> dev_bridge.py. Rename the _MissingSentinel class
to MissingSentinel, and update every importer. Also fix execution.py's relative
import to the absolute form.
