+++ tests/framework_configs/converters.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING, Any, Callable
 if TYPE_CHECKING:
     from tests.framework_configs.output_adapters import OutputAdapter
 
-from tests.framework_configs._sentinel import STRICT_CI
+from tests.framework_configs.sentinel import STRICT_CI
 
 __all__ = [
     "ConverterConfig",
